qcacld-3.0: Fix scenario for updating the country
When updating the country to be the same value, the function will return
failure. This causes a crash in the wifi start up. Add check to see if
the country code is the same to avoid calling the API.

Change-Id: I7be11857fd9cc87b65f6c598d842a9ed5c024f59
CRs-fixed: 2896855
(cherry picked from commit a052191f3bd1b5211bfbb8a24f8fae307996884c)